requirements.txt
pandas
chromadb
sentence-transformers
google-adk
litellm